requirements.txt
Django
django-propeller
django-tagulous
django-watson
